Restore project source code
Helmet v7 is going to drop support for the `Expect-CT` header. I still
plan to keep the code maintained in the `expect-ct` package, and
restoring this source code lets me do that.

Conceptually, this is a revert of
fcb65b15269da2dc767d2cc788225fb54723cf60. However, it removes the build
system and many other (development-only) dependencies.